Update variant test
