+++ property_tests/tests/func/test/test.ts
@@ -9,9 +9,7 @@ import { generateBody } from './generate_body';
 import { generateTests } from './generate_tests';
 
 const AllFuncsQueryMethod = QueryMethodArb(
-    fc.uniqueArray(FuncArb, {
-        selector: (entry) => entry.src.candidType
-    }),
+    fc.uniqueArray(FuncArb, { selector: (entry) => entry.src.candidType }),
     FuncArb,
     {
         generateBody,
